product-analyzer-v2/src/main/java/com/github/kingwaggs/productanalyzerv2/service/CategoryService.java
package com.github.kingwaggs.productanalyzerv2.service;

import com.github.kingwaggs.productanalyzerv2.domain.SelectScoreType;
import com.github.kingwaggs.productanalyzerv2.repository.CategoryRepository;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;

import java.util.List;

@Slf4j
@Service
@RequiredArgsConstructor
public class CategoryService {

    private final CategoryRepository repository;

    public List<Integer> getTargetCategoryIdList(SelectScoreType target) {
        log.info("Getting target category_id list.");
        return repository.getCategoryIdList(target);
    }

}
